package/syslog-ng: Update has for COPYING
The COPYING file in syslog-ng is not the usual GPL text. Rather, it
defines what parts of the source is licensed under what licenses; it
also contains an FAQ that is supposed to clarify how the licenses should
be interpreted for various purposes, as well as the license for a
specific subset of the soruce tree.

In 3.26.1, this file was modified according to the new or removed parts
of the source tree, but we missed that when bumping in commit ef307eefc6
(package/syslog-ng: Bump version to 2.26.1)

Update the hash accordingly.

Fixes:
 - http://autobuild.buildroot.net/results/4e0768a5aae0f77e1382108b1ca06d0ad78f4da4
 - http://autobuild.buildroot.net/results/7430550dfce0bcb1bc5c9361033e26651883685a
 - http://autobuild.buildroot.net/results/56afedff96d17ab4800845d7777dd151cb431acf
 - http://autobuild.buildroot.net/results/2cac7eeaeabf8d77724ff283d3cb9026e778635c
 - http://autobuild.buildroot.net/results/a8189eab2a14cc5d04d90d8f1fa058b86889cc21

Signed-off-by: Chris Packham <judge.packham@gmail.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>