package/network-manager: bump to version 1.22.10
Add COPYING.LGPL to license files, disable the "cloud configuration"
tool. It requires libcurl and it is too unlikely to be useful.

The library license is now LGPL-2.1+.

The hash of CONTRIBUTING has changed because it now refers to
LGPL-2.1+ instead of LGPL-2.0+.

The hash of COPYING is changed due to white-space changes, and the
removal of an introduction text:

    Unless a COPYING file in a subdirectory or file-specific license
    headers specify a different license, the following applies to all
    files in this directory and all subdirectories.

    This program is free software; you can redistribute it and/or
    modify it under the terms of the GNU General Public License as
    published by the Free Software Foundation; either version 2 of the
    License, or (at your option) any later version.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>
[Thomas:
 - fix hash of CONTRIBUTING
 - expand commit log
 - drop GFDL license information, since documentation is not installed
   on the target]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>