package/netsnmp: unix domain socket transport
The support for Unix domain socket transport was disabled as part of
the bump of netsnmp from 5.5 to 5.6 in commit
de642c9904b13914017af6f10db144d559d6bd0c, but with no apparent reason.

This support is needed to allow Unix socket based AgentX subagents to
connect to netsnmp, so let's re-enable it.

Signed-off-by: Ryan Steffens <ryan.steffens@rockwellcollins.com>
Signed-off-by: Clayton Shotwell <clayton.shotwell@rockwellcollins.com>
Signed-off-by: Matt Weber <matthew.weber@rockwellcollins.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>