configs/nanopi_neo4: new defconfig
A defconfig for the nanopi_neo4 was added in
c19128db14c08eabe319b81212d9060cf7179f95, but then removed in
326844284617cd80061cc27fbce4b9e196ed77af because an ARM32 compiler was
needed to build ATF, and this was not supported back then.

Thanks to the addition of package/arm-gnu-a-toolchain/, we can now
re-introduce this defconfig.

Compared to the previous defconfig, the following changes were done:

- enable BR2_TARGET_ARM_TRUSTED_FIRMWARE_NEEDS_ARM32_TOOLCHAIN
- bump kernel to 5.4
- increase default rootfs size to 70M
- use mainline u-boot
- switch to Marek Belisko as maintainer

Signed-off-by: Marek Belisko <marek.belisko@open-nandra.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>