configs/macchiatobin*: drop vendor defconfig, rename mainline defconfig
The support for MacchiatoBin board in U-Boot and Linux has been
significantly improved in the past years. So it no longer makes much
sense to have a separate build configuration based on older vendor
Linux kernel and U-Boot. This commit drops the configuration based on
vendor kernel and U-Boot and renames the mainline-based defconfig.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>