package/rust: remove unneded patch
The previous commit [1] added a patch to remove python2 from the configure script.
While this script is used by the Fedora rust package [2], it is not used by rust.mk.

The configure script is an alternative build system provided by upstream, and
finally calls x.py [3].

Thanks to Thomas for noticing the mistake [4].

[1] 8058c0737393f268c0d5e31129bb702c2baffe36
[2] https://src.fedoraproject.org/rpms/rust/blob/216b2d27716bf1031c526dbd0e01a1fa8e6d5aa2/f/rust.spec#_314
[3] https://github.com/rust-lang/rust/blob/1.33.0/README.md#configure-and-make
[4] http://lists.busybox.net/pipermail/buildroot/2020-May/283610.html

Signed-off-by: Romain Naour <romain.naour@gmail.com>
Cc: Titouan Christophe <titouan.christophe@railnova.eu>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>