package/redis: add optional support for TLS connections
TLS support is only enabled when BUILD_TLS is set to 'yes'. Anoy other
value, as well as empty or not set, will disable TLS. However, for
consistency with the rest of the code (e.g. for autotools packages where
we use both the --enable and --disable flags), we explcitly pass
BUILD_TLS=no. This also avoids an environment variable to leak into the
build.

Signed-off-by: Titouan Christophe <titouan.christophe@railnova.eu>
[yann.morin.1998@free.fr: add and explain the =no case]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>